Extract S3StateStoreNewTestBase
